+++ config/config.exs
@@ -1,0 +1,5 @@
+import Config
+
+config :isomer, root: File.cwd!()
+
+config :logger, level: :info